+++ app/Http/Controllers/UsuariController.php
@@ -7,6 +7,7 @@ use App\Models\Usuari;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use OpenApi\Annotations\Get;
 
 class UsuariController extends Controller
 {
@@ -83,7 +84,7 @@ class UsuariController extends Controller
             return response()->json(["Status" => "Error", "Result" => "Privilegis insuficients."], 401);
         }
 
-        $usuari = Usuari::findOrFail($id);
+        $usuari = Usuari::find($id);
         return response()->json(["Status" => "Success", "Result" => $usuari], 200);
     }
 
@@ -170,13 +171,10 @@ class UsuariController extends Controller
      *    @OA\RequestBody(
      *        required=true,
      *        @OA\JsonContent(
-     *           @OA\Property(property="ID", type="number", format="number", example="10"),
-     *           @OA\Property(property="Nom", type="string", format="string", example="Pedro"),
-     *           @OA\Property(property="Llinatges", type="string", format="string", example="Gimenez Santos"),
-     *           @OA\Property(property="Contrasenya", type="string", format="string", example="abcd1234"),
-     *           @OA\Property(property="CorreuElectronic", type="string", format="string", example="pedro@gmail.com"),
-     *           @OA\Property(property="DNI", type="string", format="string", example="12341234Q"),
-     *           @OA\Property(property="Telefon", type="string", format="string", example="971123133"),
+     *           @OA\Property(property="ID", type="number", format="number", example="24"),
+     *           @OA\Property(property="Contrasenya", type="string", format="string", example="1234"),
+     *           @OA\Property(property="CorreuElectronic", type="string", format="string", example="prova@gmail.com"),
+     *           @OA\Property(property="Telefon", type="string", format="string", example="971234567"),
      *        ),
      *     ),
      *    @OA\Response(
@@ -192,7 +190,7 @@ class UsuariController extends Controller
      *         description="Error",
      *         @OA\JsonContent(
      *         @OA\Property(property="status", type="integer", example="error"),
-     *         @OA\Property(property="data",type="string", example="Atribut DNI requerit")
+     *         @OA\Property(property="data",type="string", example="Atribut CorreuElectronic requerit")
      *         ),
      *      )
      *  )
@@ -202,12 +200,24 @@ class UsuariController extends Controller
         if ($request->ID == null || $request->ID < 1) {
             return response()->json(["Status" => "Error", "Result" => "Incorrect ID"], 400);
         }
-
         if ($request->DadesUsuari->RolsID != 3 && $request->DadesUsuari->ID != $request->ID) {
             return response()->json(["Status" => "Error", "Result" => "Privilegis insuficients."], 401);
         }
+        $usuari = Usuari::find($request->ID);
+        if ($usuari == null) {
+            return response()->json(["Status" => "Error", "Result" => "No existeix aquest usuari"], 400);
+        }
 
-        $usuari = Usuari::findOrFail($request->ID);
+        $checkEmail = Usuari::where("CorreuElectronic", "=", $request->CorreuElectronic)->get();
+        $checkTlf = Usuari::where("Telefon", "=", $request->Telefon)->get();
+
+        if (count($checkEmail) > 0 && $checkEmail[0]->ID != $request->ID) {
+            return response()->json(["Status" => "Error", "Result" => "Email ja utilitzat"], 400);
+        }
+        if (count($checkTlf) > 0 && $checkTlf[0]->ID != $request->ID) {
+            return response()->json(["Status" => "Error", "Result" => "Telefon ja utilitzat"], 400);
+        }
+
         $validator = $this->updateValidator();
         $messages = ControllersHelper::createValidatorMessages();
 
@@ -217,13 +227,10 @@ class UsuariController extends Controller
             return response()->json(["Status" => "Error", "Result" => $isValid->errors()], 400);
         }
 
-        $usuari->Nom = $request->Nom;
-        $usuari->Llinatges = $request->Llinatges;
         $usuari->Contrasenya = $request->Contrasenya;
         $usuari->CorreuElectronic = $request->CorreuElectronic;
-        $usuari->DNI = $request->DNI;
         $usuari->Telefon = $request->Telefon;
-
+        
         if ($usuari->save()) {
             return response()->json(['Status' => 'Success', 'Result' => $usuari], 200);
         } else {
@@ -276,7 +283,11 @@ class UsuariController extends Controller
             return response()->json(["Status" => "Error", "Result" => "Privilegis insuficients."], 401);
         }
 
-        $usuari = Usuari::findOrFail($request->ID);
+        $usuari = Usuari::find($request->ID);
+
+        if ($usuari == null){
+            return response()->json(["Status" => "Error", "Result" => "No existeix cap usuari amb aquesta ID"]);
+        }
 
         if ($isDeleted = $usuari->delete()) {
             return response()->json(['Status' => 'Success', 'Result' => $isDeleted], 200);
@@ -290,22 +301,20 @@ class UsuariController extends Controller
     public function createValidator(): array
     {
         return [
-            "Nom" => ["required", Rule::unique('posts', 'Nom')],
+            "Nom" => ["required"],
             "Llinatges" => ["required"],
             "Contrasenya" => ["required"],
-            "CorreuElectronic" => ["required"],
-            "DNI" => ["required"],
-            "Telefon" => ["required"]
+            "CorreuElectronic" => ["required", Rule::unique('Usuaris', 'CorreuElectronic')],
+            "DNI" => ["required", Rule::unique('Usuaris', 'DNI')],
+            "Telefon" => ["required", Rule::unique('Usuaris', 'Telefon')]
         ];
     }
     public function updateValidator(): array
     {
         return [
-            "Nom" => ["required"],
-            "Llinatges" => ["required"],
+            "ID"=> ["required"],
             "Contrasenya" => ["required"],
             "CorreuElectronic" => ["required"],
-            "DNI" => ["required"],
             "Telefon" => ["required"]
         ];
     }
